requirements.txt
typer==0.9.0
pathlib==1.0.1
numpy==1.24.1
open3d==0.17
scipy==1.10.0
pillow==9.4.0
tqdm==4.64.1
pyyaml==6.0
opencv-python==4.6.0.66
